based-on-push-and-pull_request/requirements.txt
fastapi
uvicorn
kafka-python
requests
# trufflehog3